api/requirements.txt
Django==1.9
djangorestframework==3.6.3
httpie==0.9.9
django-cors-headers==2.1.0